+++ src/CurrentProcess/Generic.php
@@ -18,22 +18,11 @@ use Innmind\Immutable\{
 final class Generic implements CurrentProcess
 {
     private Halt $halt;
-    /**
-     * @psalm-suppress DeprecatedClass
-     * @var Set<Child>
-     */
-    private Set $children;
-    private ?Handler $signalsHandler = null;
     private ?Signals\Wrapper $signals = null;
 
     private function __construct(Halt $halt)
     {
         $this->halt = $halt;
-        /**
-         * @psalm-suppress DeprecatedClass
-         * @var Set<Child>
-         */
-        $this->children = Set::of();
     }
 
     public static function of(Halt $halt): self
@@ -47,50 +36,9 @@ final class Generic implements CurrentProcess
         return new Pid(\getmypid());
     }
 
-    public function fork(): Either
-    {
-        /**
-         * @psalm-suppress ArgumentTypeCoercion
-         * @psalm-suppress DeprecatedClass
-         * @var Either<ForkFailed|Child, SideEffect>
-         */
-        $result = match ($pid = \pcntl_fork()) {
-            -1 => Either::left(new ForkFailed),
-            0 => Either::right(new SideEffect),
-            default => Either::left(Child::of(new Pid($pid))),
-        };
-
-        /**
-         * @psalm-suppress DeprecatedClass
-         */
-        return $result
-            ->map(function($sideEffect) {
-                $this->children = $this->children->clear();
-                $this->signals = null;
-                $this->signalsHandler = $this->signalsHandler ? $this->signalsHandler->reset() : null;
-
-                return $sideEffect;
-            })
-            ->leftMap(fn($left) => match (true) {
-                $left instanceof Child => $this->register($left),
-                default => $left,
-            });
-    }
-
-    /**
-     * @psalm-suppress DeprecatedClass
-     */
-    public function children(): Children
-    {
-        /** @psalm-suppress DeprecatedClass */
-        return Children::of(...$this->children->toList());
-    }
-
     public function signals(): Signals
     {
-        return $this->signals ??= Signals\Wrapper::of(
-            $this->signalsHandler = new Handler,
-        );
+        return $this->signals ??= Signals\Wrapper::of(new Handler);
     }
 
     public function halt(Period $period): void
@@ -101,15 +49,5 @@ final class Generic implements CurrentProcess
     public function memory(): Bytes
     {
         return new Bytes(\memory_get_usage());
-    }
-
-    /**
-     * @psalm-suppress DeprecatedClass
-     */
-    private function register(Child $child): Child
-    {
-        $this->children = ($this->children)($child);
-
-        return $child;
     }
 }