+++ src/CurrentProcess/Logger.php
@@ -7,7 +7,6 @@ use Innmind\OperatingSystem\CurrentProcess;
 use Innmind\Server\Control\Server\Process\Pid;
 use Innmind\Server\Status\Server\Memory\Bytes;
 use Innmind\TimeContinuum\Period;
-use Innmind\Immutable\Either;
 use Psr\Log\LoggerInterface;
 
 final class Logger implements CurrentProcess
@@ -37,38 +36,6 @@ final class Logger implements CurrentProcess
         );
 
         return $pid;
-    }
-
-    public function fork(): Either
-    {
-        $this->logger->debug('Forking process');
-
-        /**
-         * @psalm-suppress DeprecatedMethod
-         * @psalm-suppress DeprecatedClass
-         */
-        return $this
-            ->process
-            ->fork()
-            ->map(function($sideEffect) {
-                // @see Generic::fork()
-                // the child process reset the signal handlers to avoid side effects
-                // between parent and child so here we can safely reset the signals
-                // to free memory as listeners are kept in memory to reference the
-                // decorated listeners given to the underlying signals handler
-                $this->signals = null;
-
-                return $sideEffect;
-            });
-    }
-
-    /**
-     * @psalm-suppress DeprecatedClass
-     */
-    public function children(): Children
-    {
-        /** @psalm-suppress DeprecatedMethod */
-        return $this->process->children();
     }
 
     public function signals(): Signals